Fix emergency_close warning
`ErlDrvEntry` got a new field, `emergency_close` in 17.4. This corresponds
to `ERL_DRV_EXTENDED_MAJOR_VERSION=3` and `ERL_DRV_EXTENDED_MINOR_VERSION=2`.
This patch adds a check to see if the version is >= 3.2 and adds an
initialized for `emergency_close` if it's needed. This gets rid of an annoying
warning (missing initializer) for recent Erlang versions.